fix: restructure to fix managed support
