+++ package.json
@@ -5,8 +5,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
-    "dist",
-    "dist/upload-sourcemaps.js"
+    "dist"
   ],
   "license": "MIT",
   "keywords": [
