+++ src/integrations/managed.ts
@@ -1,0 +1,94 @@
+import { Platform } from 'react-native';
+import Constants from 'expo-constants';
+import * as Device from 'expo-device';
+import {
+  setExtras,
+  setTags,
+  getCurrentHub,
+  Severity,
+  setTag,
+  addGlobalEventProcessor,
+} from '@sentry/react-native';
+
+export class ExpoIntegration {
+  static id = 'ExpoIntegration';
+  name = ExpoIntegration.id;
+
+  setupOnce() {
+    let manifest = Constants.manifest;
+    setExtras({
+      manifest: manifest,
+      deviceYearClass: Constants.deviceYearClass,
+      linkingUri: Constants.linkingUri,
+    });
+
+    setTags({
+      deviceId: Constants.installationId,
+      appOwnership: Constants.appOwnership,
+    });
+
+    if (Constants.appOwnership === 'expo' && Constants.expoVersion) {
+      setTag('expoAppVersion', Constants.expoVersion);
+    }
+
+    if (!!manifest) {
+      setTag('expoReleaseChannel', manifest.releaseChannel);
+      setTag('appVersion', manifest.version ?? '');
+      setTag('appPublishedTime', manifest.publishedTime);
+      setTag('expoSdkVersion', manifest.sdkVersion ?? '');
+    }
+
+    const defaultHandler = ErrorUtils.getGlobalHandler();
+
+    ErrorUtils.setGlobalHandler((error, isFatal) => {
+      // On Android, the Expo bundle filepath cannot be handled by TraceKit,
+      // so we normalize it to use the same filepath that we use on Expo iOS.
+      if (Platform.OS === 'android') {
+        error.stack = error.stack.replace(
+          /\/.*\/\d+\.\d+.\d+\/cached\-bundle\-experience\-/g,
+          'https://d1wp6m56sqw74a.cloudfront.net:443/'
+        );
+      }
+
+      getCurrentHub().withScope((scope) => {
+        if (isFatal) {
+          scope.setLevel(Severity.Fatal);
+        }
+        getCurrentHub().captureException(error, {
+          originalException: error,
+        });
+      });
+
+      const client = getCurrentHub().getClient();
+      if (client && !__DEV__) {
+        // @ts-ignore PR to add this to types: https://github.com/getsentry/sentry-javascript/pull/2669
+        client.flush(client.getOptions().shutdownTimeout || 2000).then(() => {
+          defaultHandler(error, isFatal);
+        });
+      } else {
+        // If there is no client, something is fishy but we call the default handler anyway. Even if in dev
+        defaultHandler(error, isFatal);
+      }
+    });
+
+    addGlobalEventProcessor(function (event, _hint) {
+      const that = getCurrentHub().getIntegration(ExpoIntegration);
+
+      if (that) {
+        event.contexts = {
+          ...(event.contexts || {}),
+          device: {
+            simulator: !Device.isDevice,
+            model: Device.modelName,
+          },
+          os: {
+            name: Device.osName,
+            version: Device.osVersion,
+          },
+        };
+      }
+
+      return event;
+    });
+  }
+}
